fix(index.js): changes in pipeline, file structure, package and dependencies
