--- .babelrc
@@ -1,5 +1,0 @@
-{
-    "comments": false,
-    "presets": ["@babel/preset-env"],
-    "plugins": ["@babel/plugin-proposal-class-properties", "@babel/plugin-proposal-object-rest-spread"]
-}
