+++ .github/workflows/nodejs.yml
@@ -1,7 +1,7 @@
 # This workflow will do a clean installation of node dependencies, cache/restore them, build the source code and run tests across different versions of node
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-nodejs-with-github-actions
 
-name: Node.js CI
+name: Node.js CI Build
 
 on:
   push:
@@ -35,5 +35,7 @@ jobs:
     - name: Install Dependencies
       if: steps.cache-nodemodules.outputs.cache-hint != 'true'
       run: npm ci
+    - name: Run Linter
+      run: npm lint
     - name: Test app
       run: npm test
