+++ index.js
@@ -1,240 +1,190 @@
-"use strict";
+/**
+ * easy-version module
+ * Middleware to work with API Versioning
+ *
+ * - registerVersion: Register API version
+ * For detect the version requested in Accept headers and set to the request object into the
+ * version attribute.
+ * The version must be using the format specified in the RFC https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html
+ * The version must be specify using the keywords "version" or "v"
+ * E.g.
+ *     Accept: application/x.my-microservice-name+json;version=1.0.0
+ *     Accept: application/x.my-microservice-name+json;v=1.0.0
+ *
+ * - routeTo: Executes the handler
+ * For executes the correct handler for the version requested
+ * The args parameter its an object that maps the version to the handler.
+ * E.g.
+ *      {
+ *          "1.0.0":handlerV1,
+ *          "2.0.0":handlerV2
+ *      }
+ * @author Rafael Chavez Solis <rafaelchavezsolis@gmail.com>
 
-function _regeneratorRuntime() { "use strict"; _regeneratorRuntime = function _regeneratorRuntime() { return exports; }; var exports = {}, Op = Object.prototype, hasOwn = Op.hasOwnProperty, $Symbol = "function" == typeof Symbol ? Symbol : {}, iteratorSymbol = $Symbol.iterator || "@@iterator", asyncIteratorSymbol = $Symbol.asyncIterator || "@@asyncIterator", toStringTagSymbol = $Symbol.toStringTag || "@@toStringTag"; function define(obj, key, value) { return Object.defineProperty(obj, key, { value: value, enumerable: !0, configurable: !0, writable: !0 }), obj[key]; } try { define({}, ""); } catch (err) { define = function define(obj, key, value) { return obj[key] = value; }; } function wrap(innerFn, outerFn, self, tryLocsList) { var protoGenerator = outerFn && outerFn.prototype instanceof Generator ? outerFn : Generator, generator = Object.create(protoGenerator.prototype), context = new Context(tryLocsList || []); return generator._invoke = function (innerFn, self, context) { var state = "suspendedStart"; return function (method, arg) { if ("executing" === state) throw new Error("Generator is already running"); if ("completed" === state) { if ("throw" === method) throw arg; return doneResult(); } for (context.method = method, context.arg = arg;;) { var delegate = context.delegate; if (delegate) { var delegateResult = maybeInvokeDelegate(delegate, context); if (delegateResult) { if (delegateResult === ContinueSentinel) continue; return delegateResult; } } if ("next" === context.method) context.sent = context._sent = context.arg;else if ("throw" === context.method) { if ("suspendedStart" === state) throw state = "completed", context.arg; context.dispatchException(context.arg); } else "return" === context.method && context.abrupt("return", context.arg); state = "executing"; var record = tryCatch(innerFn, self, context); if ("normal" === record.type) { if (state = context.done ? "completed" : "suspendedYield", record.arg === ContinueSentinel) continue; return { value: record.arg, done: context.done }; } "throw" === record.type && (state = "completed", context.method = "throw", context.arg = record.arg); } }; }(innerFn, self, context), generator; } function tryCatch(fn, obj, arg) { try { return { type: "normal", arg: fn.call(obj, arg) }; } catch (err) { return { type: "throw", arg: err }; } } exports.wrap = wrap; var ContinueSentinel = {}; function Generator() {} function GeneratorFunction() {} function GeneratorFunctionPrototype() {} var IteratorPrototype = {}; define(IteratorPrototype, iteratorSymbol, function () { return this; }); var getProto = Object.getPrototypeOf, NativeIteratorPrototype = getProto && getProto(getProto(values([]))); NativeIteratorPrototype && NativeIteratorPrototype !== Op && hasOwn.call(NativeIteratorPrototype, iteratorSymbol) && (IteratorPrototype = NativeIteratorPrototype); var Gp = GeneratorFunctionPrototype.prototype = Generator.prototype = Object.create(IteratorPrototype); function defineIteratorMethods(prototype) { ["next", "throw", "return"].forEach(function (method) { define(prototype, method, function (arg) { return this._invoke(method, arg); }); }); } function AsyncIterator(generator, PromiseImpl) { function invoke(method, arg, resolve, reject) { var record = tryCatch(generator[method], generator, arg); if ("throw" !== record.type) { var result = record.arg, value = result.value; return value && "object" == _typeof(value) && hasOwn.call(value, "__await") ? PromiseImpl.resolve(value.__await).then(function (value) { invoke("next", value, resolve, reject); }, function (err) { invoke("throw", err, resolve, reject); }) : PromiseImpl.resolve(value).then(function (unwrapped) { result.value = unwrapped, resolve(result); }, function (error) { return invoke("throw", error, resolve, reject); }); } reject(record.arg); } var previousPromise; this._invoke = function (method, arg) { function callInvokeWithMethodAndArg() { return new PromiseImpl(function (resolve, reject) { invoke(method, arg, resolve, reject); }); } return previousPromise = previousPromise ? previousPromise.then(callInvokeWithMethodAndArg, callInvokeWithMethodAndArg) : callInvokeWithMethodAndArg(); }; } function maybeInvokeDelegate(delegate, context) { var method = delegate.iterator[context.method]; if (undefined === method) { if (context.delegate = null, "throw" === context.method) { if (delegate.iterator["return"] && (context.method = "return", context.arg = undefined, maybeInvokeDelegate(delegate, context), "throw" === context.method)) return ContinueSentinel; context.method = "throw", context.arg = new TypeError("The iterator does not provide a 'throw' method"); } return ContinueSentinel; } var record = tryCatch(method, delegate.iterator, context.arg); if ("throw" === record.type) return context.method = "throw", context.arg = record.arg, context.delegate = null, ContinueSentinel; var info = record.arg; return info ? info.done ? (context[delegate.resultName] = info.value, context.next = delegate.nextLoc, "return" !== context.method && (context.method = "next", context.arg = undefined), context.delegate = null, ContinueSentinel) : info : (context.method = "throw", context.arg = new TypeError("iterator result is not an object"), context.delegate = null, ContinueSentinel); } function pushTryEntry(locs) { var entry = { tryLoc: locs[0] }; 1 in locs && (entry.catchLoc = locs[1]), 2 in locs && (entry.finallyLoc = locs[2], entry.afterLoc = locs[3]), this.tryEntries.push(entry); } function resetTryEntry(entry) { var record = entry.completion || {}; record.type = "normal", delete record.arg, entry.completion = record; } function Context(tryLocsList) { this.tryEntries = [{ tryLoc: "root" }], tryLocsList.forEach(pushTryEntry, this), this.reset(!0); } function values(iterable) { if (iterable) { var iteratorMethod = iterable[iteratorSymbol]; if (iteratorMethod) return iteratorMethod.call(iterable); if ("function" == typeof iterable.next) return iterable; if (!isNaN(iterable.length)) { var i = -1, next = function next() { for (; ++i < iterable.length;) { if (hasOwn.call(iterable, i)) return next.value = iterable[i], next.done = !1, next; } return next.value = undefined, next.done = !0, next; }; return next.next = next; } } return { next: doneResult }; } function doneResult() { return { value: undefined, done: !0 }; } return GeneratorFunction.prototype = GeneratorFunctionPrototype, define(Gp, "constructor", GeneratorFunctionPrototype), define(GeneratorFunctionPrototype, "constructor", GeneratorFunction), GeneratorFunction.displayName = define(GeneratorFunctionPrototype, toStringTagSymbol, "GeneratorFunction"), exports.isGeneratorFunction = function (genFun) { var ctor = "function" == typeof genFun && genFun.constructor; return !!ctor && (ctor === GeneratorFunction || "GeneratorFunction" === (ctor.displayName || ctor.name)); }, exports.mark = function (genFun) { return Object.setPrototypeOf ? Object.setPrototypeOf(genFun, GeneratorFunctionPrototype) : (genFun.__proto__ = GeneratorFunctionPrototype, define(genFun, toStringTagSymbol, "GeneratorFunction")), genFun.prototype = Object.create(Gp), genFun; }, exports.awrap = function (arg) { return { __await: arg }; }, defineIteratorMethods(AsyncIterator.prototype), define(AsyncIterator.prototype, asyncIteratorSymbol, function () { return this; }), exports.AsyncIterator = AsyncIterator, exports.async = function (innerFn, outerFn, self, tryLocsList, PromiseImpl) { void 0 === PromiseImpl && (PromiseImpl = Promise); var iter = new AsyncIterator(wrap(innerFn, outerFn, self, tryLocsList), PromiseImpl); return exports.isGeneratorFunction(outerFn) ? iter : iter.next().then(function (result) { return result.done ? result.value : iter.next(); }); }, defineIteratorMethods(Gp), define(Gp, toStringTagSymbol, "Generator"), define(Gp, iteratorSymbol, function () { return this; }), define(Gp, "toString", function () { return "[object Generator]"; }), exports.keys = function (object) { var keys = []; for (var key in object) { keys.push(key); } return keys.reverse(), function next() { for (; keys.length;) { var key = keys.pop(); if (key in object) return next.value = key, next.done = !1, next; } return next.done = !0, next; }; }, exports.values = values, Context.prototype = { constructor: Context, reset: function reset(skipTempReset) { if (this.prev = 0, this.next = 0, this.sent = this._sent = undefined, this.done = !1, this.delegate = null, this.method = "next", this.arg = undefined, this.tryEntries.forEach(resetTryEntry), !skipTempReset) for (var name in this) { "t" === name.charAt(0) && hasOwn.call(this, name) && !isNaN(+name.slice(1)) && (this[name] = undefined); } }, stop: function stop() { this.done = !0; var rootRecord = this.tryEntries[0].completion; if ("throw" === rootRecord.type) throw rootRecord.arg; return this.rval; }, dispatchException: function dispatchException(exception) { if (this.done) throw exception; var context = this; function handle(loc, caught) { return record.type = "throw", record.arg = exception, context.next = loc, caught && (context.method = "next", context.arg = undefined), !!caught; } for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i], record = entry.completion; if ("root" === entry.tryLoc) return handle("end"); if (entry.tryLoc <= this.prev) { var hasCatch = hasOwn.call(entry, "catchLoc"), hasFinally = hasOwn.call(entry, "finallyLoc"); if (hasCatch && hasFinally) { if (this.prev < entry.catchLoc) return handle(entry.catchLoc, !0); if (this.prev < entry.finallyLoc) return handle(entry.finallyLoc); } else if (hasCatch) { if (this.prev < entry.catchLoc) return handle(entry.catchLoc, !0); } else { if (!hasFinally) throw new Error("try statement without catch or finally"); if (this.prev < entry.finallyLoc) return handle(entry.finallyLoc); } } } }, abrupt: function abrupt(type, arg) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.tryLoc <= this.prev && hasOwn.call(entry, "finallyLoc") && this.prev < entry.finallyLoc) { var finallyEntry = entry; break; } } finallyEntry && ("break" === type || "continue" === type) && finallyEntry.tryLoc <= arg && arg <= finallyEntry.finallyLoc && (finallyEntry = null); var record = finallyEntry ? finallyEntry.completion : {}; return record.type = type, record.arg = arg, finallyEntry ? (this.method = "next", this.next = finallyEntry.finallyLoc, ContinueSentinel) : this.complete(record); }, complete: function complete(record, afterLoc) { if ("throw" === record.type) throw record.arg; return "break" === record.type || "continue" === record.type ? this.next = record.arg : "return" === record.type ? (this.rval = this.arg = record.arg, this.method = "return", this.next = "end") : "normal" === record.type && afterLoc && (this.next = afterLoc), ContinueSentinel; }, finish: function finish(finallyLoc) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.finallyLoc === finallyLoc) return this.complete(entry.completion, entry.afterLoc), resetTryEntry(entry), ContinueSentinel; } }, "catch": function _catch(tryLoc) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.tryLoc === tryLoc) { var record = entry.completion; if ("throw" === record.type) { var thrown = record.arg; resetTryEntry(entry); } return thrown; } } throw new Error("illegal catch attempt"); }, delegateYield: function delegateYield(iterable, resultName, nextLoc) { return this.delegate = { iterator: values(iterable), resultName: resultName, nextLoc: nextLoc }, "next" === this.method && (this.arg = undefined), ContinueSentinel; } }, exports; }
-
-function asyncGeneratorStep(gen, resolve, reject, _next, _throw, key, arg) { try { var info = gen[key](arg); var value = info.value; } catch (error) { reject(error); return; } if (info.done) { resolve(value); } else { Promise.resolve(value).then(_next, _throw); } }
-
-function _asyncToGenerator(fn) { return function () { var self = this, args = arguments; return new Promise(function (resolve, reject) { var gen = fn.apply(self, args); function _next(value) { asyncGeneratorStep(gen, resolve, reject, _next, _throw, "next", value); } function _throw(err) { asyncGeneratorStep(gen, resolve, reject, _next, _throw, "throw", err); } _next(undefined); }); }; }
-
-function _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }
-
-function _nonIterableRest() { throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }
-
-function _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== "undefined" && arr[Symbol.iterator] || arr["@@iterator"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i["return"] != null) _i["return"](); } finally { if (_d) throw _e; } } return _arr; }
-
-function _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }
-
-function _createForOfIteratorHelper(o, allowArrayLike) { var it = typeof Symbol !== "undefined" && o[Symbol.iterator] || o["@@iterator"]; if (!it) { if (Array.isArray(o) || (it = _unsupportedIterableToArray(o)) || allowArrayLike && o && typeof o.length === "number") { if (it) o = it; var i = 0; var F = function F() {}; return { s: F, n: function n() { if (i >= o.length) return { done: true }; return { done: false, value: o[i++] }; }, e: function e(_e2) { throw _e2; }, f: F }; } throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); } var normalCompletion = true, didErr = false, err; return { s: function s() { it = it.call(o); }, n: function n() { var step = it.next(); normalCompletion = step.done; return step; }, e: function e(_e3) { didErr = true; err = _e3; }, f: function f() { try { if (!normalCompletion && it["return"] != null) it["return"](); } finally { if (didErr) throw err; } } }; }
-
-function _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === "string") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === "Object" && o.constructor) n = o.constructor.name; if (n === "Map" || n === "Set") return Array.from(o); if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }
-
-function _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }
-
-function _typeof(obj) { "@babel/helpers - typeof"; return _typeof = "function" == typeof Symbol && "symbol" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && "function" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }, _typeof(obj); }
-
-var errors = {
-  missingHeader: 'Accept header is required.',
-  customFunctionError: 'An error occurs while parsing version using custom function.',
-  missingVersion: 'You must specify the version number.',
-  invalidVersionNumber: 'You must provide a valid version number.',
-  invalidRoutes: 'You must provide a valid route handler object map.',
-  handlerError: 'You must provide a valid handler for the version requested.'
+/**
+ * Error messages
+ */
+const errors = {
+    missingHeader: 'Accept header is required.',
+    customFunctionError: 'An error occurs while parsing version using custom function.',
+    missingVersion: 'You must specify the version number.',
+    invalidVersionNumber: 'You must provide a valid version number.',
+    invalidRoutes: 'You must provide a valid route handler object map.',
+    handlerError: 'You must provide a valid handler for the version requested.',
 };
-var MAX_VERSION_DIGITS = 3;
-
-var removeWhitespaces = function removeWhitespaces(str) {
-  if (typeof str === 'string') {
-    return str.replace(/\s/g, '');
-  }
-
-  return '';
+const MAX_VERSION_DIGITS = 3;
+/**
+ * Function to remove white spaces
+ * @param {string} str Generic string
+ * @returns string with no white spaces
+ */
+const removeWhitespaces = (str) => {
+    if (typeof str === 'string') {
+        return str.replace(/\s/g, '');
+    }
+    return '';
 };
-
-var formatVersion = function formatVersion(version) {
-  if (!version || typeof version === 'function' || _typeof(version) === 'object' || version === true) {
-    return undefined;
-  }
-
-  var ver = removeWhitespaces(version.toString());
-  var split = ver.split('.');
-
-  if (split.length < MAX_VERSION_DIGITS) {
-    var size;
-
-    if (split.length < 0) {
-      size = 0;
+/**
+ * Function to standarize the version number to a 3 digits length
+ * @param {string} version Value to format
+ * @returns Version standarized to a 3 digits length
+ */
+const formatVersion = (version) => {
+    if (!version || typeof version === 'function' || typeof version === 'object' || version === true) {
+        return undefined;
+    }
+    let ver = removeWhitespaces(version.toString());
+    const split = ver.split('.');
+    if (split.length < MAX_VERSION_DIGITS) {
+        let size;
+        if (split.length < 0) {
+            size = 0;
+        } else {
+            size = split.length;
+        }
+        for (let i = size; i < MAX_VERSION_DIGITS; i++) {
+            ver += '.0';
+        }
+        return ver;
+    } else if (split.length > MAX_VERSION_DIGITS) {
+        return split.slice(0, MAX_VERSION_DIGITS).join('.');
     } else {
-      size = split.length;
+        return ver;
     }
-
-    for (var i = size; i < MAX_VERSION_DIGITS; i++) {
-      ver += '.0';
-    }
-
-    return ver;
-  } else if (split.length > MAX_VERSION_DIGITS) {
-    return split.slice(0, MAX_VERSION_DIGITS).join('.');
-  } else {
-    return ver;
-  }
 };
+/**
+ * Function to parse the accept header value to find the version requested and assign it to request version attribute
+ * @param {function} customFunction Function to parse version
+ */
 
 function registerVersion(customFunction) {
-  return function (req, res, next) {
-    if (!req || !req.headers || !req.headers.accept || req.headers.accept === undefined || typeof req.headers.accept !== 'string') {
-      next(new Error(errors.missingHeader));
-    } else {
-      if (customFunction && typeof customFunction === 'function') {
-        try {
-          req.version = customFunction(req.get('accept'));
-          next();
-        } catch (err) {
-          next(new Error("".concat(errors.customFunctionError, " ").concat(err.toString())));
-        }
-      } else {
-        var acceptHeader = removeWhitespaces(req.headers.accept).replace(/\r\n\S.*$/, '');
-        acceptHeader = acceptHeader.length > 0 && acceptHeader.split(';');
-
-        if (!acceptHeader || acceptHeader === undefined || acceptHeader.length !== 2) {
-          next(new Error(errors.missingVersion));
+    return function (req, res, next) {
+        // If no exits the accept header
+        if (
+            !req ||
+            !req.headers ||
+            !req.headers.accept ||
+            req.headers.accept === undefined ||
+            typeof req.headers.accept !== 'string'
+        ) {
+            next(new Error(errors.missingHeader));
         } else {
-          var versionNumber;
-          var params = acceptHeader[1].split(',');
-
-          if (params !== undefined && params.length > 0) {
-            var _iterator = _createForOfIteratorHelper(params),
-                _step;
-
-            try {
-              for (_iterator.s(); !(_step = _iterator.n()).done;) {
-                var param = _step.value;
-
-                var _param$split = param.split('='),
-                    _param$split2 = _slicedToArray(_param$split, 2),
-                    key = _param$split2[0],
-                    value = _param$split2[1];
-
-                if (key === 'version' || key === 'v') {
-                  versionNumber = formatVersion(value);
-                  break;
+            // If received a custom function to parse the version header
+            if (customFunction && typeof customFunction === 'function') {
+                try {
+                    // Call the custom function
+                    req.version = customFunction(req.get('accept'));
+                    next();
+                } catch (err) {
+                    next(new Error(`${errors.customFunctionError} ${err.toString()}`));
                 }
-              }
-            } catch (err) {
-              _iterator.e(err);
-            } finally {
-              _iterator.f();
+            } else {
+                // Get header value
+                let acceptHeader = removeWhitespaces(req.headers.accept).replace(/\r\n\S.*$/, '');
+                acceptHeader = acceptHeader.length > 0 && acceptHeader.split(';');
+                // If not receive params
+                if (!acceptHeader || acceptHeader === undefined || acceptHeader.length !== 2) {
+                    next(new Error(errors.missingVersion));
+                } else {
+                    let versionNumber;
+                    const params = acceptHeader[1].split(',');
+                    // For each param, search for version
+                    if (params !== undefined && params.length > 0) {
+                        for (const param of params) {
+                            const [key, value] = param.split('=');
+                            if (key === 'version' || key === 'v') {
+                                // Parse version
+                                versionNumber = formatVersion(value);
+                                break;
+                            }
+                        }
+                    }
+                    // If version detected
+                    if (versionNumber !== undefined) {
+                        // Sets the version
+                        req.version = versionNumber;
+                        // Set common response headers
+                        res.setHeader('Content-Type', `${acceptHeader[0]};version=${versionNumber}`);
+                        res.setHeader('Vary', 'Content-Type');
+                        // Move to the next middleware
+                        next();
+                    } else {
+                        next(new Error(errors.invalidVersionNumber));
+                    }
+                }
             }
-          }
-
-          if (versionNumber !== undefined) {
-            req.version = versionNumber;
-            res.setHeader('Content-Type', "".concat(acceptHeader[0], ";version=").concat(versionNumber));
-            res.setHeader('Vary', 'Content-Type');
-            next();
-          } else {
-            next(new Error(errors.invalidVersionNumber));
-          }
         }
-      }
-    }
-  };
-}
-
-function routeTo(args, notVersionFound) {
-  return function () {
-    var _ref = _asyncToGenerator(_regeneratorRuntime().mark(function _callee(req, res, next) {
-      var versions, rv, v, _i2, _versions, version;
-
-      return _regeneratorRuntime().wrap(function _callee$(_context) {
-        while (1) {
-          switch (_context.prev = _context.next) {
-            case 0:
-              _context.prev = 0;
-
-              if (!(!req || !req.version || typeof req.version !== 'string')) {
-                _context.next = 4;
-                break;
-              }
-
-              res.statusCode = 403;
-              throw new Error(errors.missingVersion);
-
-            case 4:
-              if (!(!args || _typeof(args) !== 'object' || Array.isArray(args))) {
-                _context.next = 7;
-                break;
-              }
-
-              res.statusCode = 403;
-              throw new Error(errors.invalidRoutes);
-
-            case 7:
-              versions = Object.keys(args);
-              rv = String(req.version);
-              _i2 = 0, _versions = versions;
-
-            case 10:
-              if (!(_i2 < _versions.length)) {
-                _context.next = 19;
-                break;
-              }
-
-              version = _versions[_i2];
-              version = String(version);
-
-              if (!(version.toLowerCase().trim() === rv.toLowerCase().trim())) {
-                _context.next = 16;
-                break;
-              }
-
-              v = version;
-              return _context.abrupt("break", 19);
-
-            case 16:
-              _i2++;
-              _context.next = 10;
-              break;
-
-            case 19:
-              if (!(v === undefined)) {
-                _context.next = 27;
-                break;
-              }
-
-              if (notVersionFound) {
-                _context.next = 23;
-                break;
-              }
-
-              res.statusCode = 403;
-              throw new Error(errors.handlerError);
-
-            case 23:
-              _context.next = 25;
-              return notVersionFound.call(this, req, res, next);
-
-            case 25:
-              _context.next = 29;
-              break;
-
-            case 27:
-              _context.next = 29;
-              return args[v].call(this, req, res, next);
-
-            case 29:
-              _context.next = 34;
-              break;
-
-            case 31:
-              _context.prev = 31;
-              _context.t0 = _context["catch"](0);
-              next(_context.t0);
-
-            case 34:
-            case "end":
-              return _context.stop();
-          }
-        }
-      }, _callee, this, [[0, 31]]);
-    }));
-
-    return function (_x, _x2, _x3) {
-      return _ref.apply(this, arguments);
     };
-  }();
 }
-
+/**
+ * Function to execute the correct handler of the version requested
+ * @param {object} args Map with the version number and its handler
+ * @param {function} notVersionFound Default callback for executes if no version found
+ */
+function routeTo(args, notVersionFound) {
+    return async function (req, res, next) {
+        try {
+            // If no exits the accept header
+            if (!req || !req.version || typeof req.version !== 'string') {
+                res.statusCode = 403;
+                throw new Error(errors.missingVersion);
+            }
+            // If not received a map for routing
+            if (!args || typeof args !== 'object' || Array.isArray(args)) {
+                res.statusCode = 403;
+                throw new Error(errors.invalidRoutes);
+            }
+            // Gets map and execute the handler
+            const versions = Object.keys(args);
+            const rv = String(req.version);
+            let v;
+            for (let version of versions) {
+                version = String(version);
+                if (version.toLowerCase().trim() === rv.toLowerCase().trim()) {
+                    v = version;
+                    break;
+                }
+            }
+            // If not version found
+            if (v === undefined) {
+                // Executes callback if no version its found
+                if (!notVersionFound) {
+                    res.statusCode = 403;
+                    throw new Error(errors.handlerError);
+                }
+                await notVersionFound.call(this, req, res, next);
+            } else {
+                await args[v].call(this, req, res, next);
+            }
+        } catch (err) {
+            next(err);
+        }
+    };
+}
+/**
+ * Exports module
+ */
 module.exports = {
-  registerVersion: registerVersion,
-  routeTo: routeTo
+    registerVersion,
+    routeTo,
 };
